Change component state hierarchy and add slight UI improvements
