--- components/EditorHeader/index.ts
@@ -1,5 +1,0 @@
-import { EditorHeader } from './Header'
-import { EditorCover } from './Cover'
-import { EditorIcon } from './Icon'
-
-export { EditorHeader, EditorCover, EditorIcon }
